guf Yeah!

+++ src/Kebler.Update/MainWindow.xaml.cs
@@ -7,17 +7,19 @@ using System.IO.Compression;
 using System.Net;
 using System.Net.Mime;
 using System.Windows;
+using Kebler.Const;
 
 namespace Kebler.Update
 {
     /// <summary>
-    /// Interaction logic for MainWindow.xaml
+    ///     Interaction logic for MainWindow.xaml
     /// </summary>
     public partial class MainWindow : Window
     {
-        Uri uri;
+        private readonly Uri uri;
         private MyWebClient _webClient;
         private string tempfile;
+
         public MainWindow(Uri uri)
         {
             InitializeComponent();
@@ -40,32 +42,24 @@ namespace Kebler.Update
 
         private void WebClientOnDownloadFileCompleted(object sender, AsyncCompletedEventArgs e)
         {
-            if (e.Cancelled)
-            {
-                return;
-            }
+            if (e.Cancelled) return;
 
             ContentDisposition contentDisposition = null;
             if (_webClient.ResponseHeaders?["Content-Disposition"] != null)
-            {
                 contentDisposition = new ContentDisposition(_webClient.ResponseHeaders["Content-Disposition"]);
-            }
 
             var fileName = contentDisposition.FileName;
 
             var pth = Path.Combine(Path.GetTempPath(), fileName);
-            if (File.Exists(pth))
-            {
-                File.Delete(pth);
-            }
+            if (File.Exists(pth)) File.Delete(pth);
             File.Move(tempfile, pth);
 
             var zip = new ZipArchive(new FileStream(pth, FileMode.Open));
-            zip.ExtractToDirectory(Const.ConstStrings.KeblerRoamingFolder, true);
+            zip.ExtractToDirectory(ConstStrings.KeblerRoamingFolder, true);
 
             var processStartInfo = new ProcessStartInfo
             {
-                FileName = Const.ConstStrings.KeblerExepath,
+                FileName = ConstStrings.KeblerExepath
             };
             App.Instance.CreateShortcut();
             Process.Start(processStartInfo);
@@ -74,10 +68,7 @@ namespace Kebler.Update
 
         public static void DeleteDirectory(string path)
         {
-            foreach (var directory in Directory.GetDirectories(path))
-            {
-                DeleteDirectory(directory);
-            }
+            foreach (var directory in Directory.GetDirectories(path)) DeleteDirectory(directory);
 
             try
             {
@@ -116,7 +107,7 @@ namespace Kebler.Update
 
         private static string BytesToString(long byteCount)
         {
-            string[] suf = { "B", "KB", "MB", "GB", "TB", "PB", "EB" };
+            string[] suf = {"B", "KB", "MB", "GB", "TB", "PB", "EB"};
             if (byteCount == 0)
                 return "0" + suf[0];
             var bytes = Math.Abs(byteCount);
@@ -124,7 +115,6 @@ namespace Kebler.Update
             var num = Math.Round(bytes / Math.Pow(1024, place), 1);
             return $"{(Math.Sign(byteCount) * num).ToString(CultureInfo.InvariantCulture)} {suf[place]}";
         }
-
     }
 
     public class MyWebClient : WebClient
